Move files around and add submodule
