+++ notepad-to-remove/DockingFeature/dockingResource.h
