+++ notepad-to-remove/PluginInterface.h
