+++ notepad-to-remove/Scintilla.h
