--- resource.h
@@ -1,20 +1,0 @@
-//{{NO_DEPENDENCIES}}
-// Microsoft Visual C++ generated include file.
-// Used by Resource.rc
-//
-#define IDD_OUTPUT                      100
-#define IDI_ICON1                       103
-#define IDC_TABBAR                      1001
-#define IDC_LIST_OUTPUT                 1002
-#define IDC_LIST_LINTS                  1003
-
-// Next default values for new objects
-// 
-#ifdef APSTUDIO_INVOKED
-#ifndef APSTUDIO_READONLY_SYMBOLS
-#define _APS_NEXT_RESOURCE_VALUE        105
-#define _APS_NEXT_COMMAND_VALUE         40001
-#define _APS_NEXT_CONTROL_VALUE         1004
-#define _APS_NEXT_SYMED_VALUE           101
-#endif
-#endif
