+++ src/Clipboard.cpp
