+++ src/Clipboard.h
