+++ src/XmlDecodeException.h
