+++ src/encoding.cpp
