+++ src/encoding.h
